--- example/dubbo/go-client/conf/client.yml
@@ -1,22 +1,0 @@
-# dubbo client yaml configure file
-# application config (not necessary)
-application:
-  organization: "dubbo.io"
-  name: "GreeterProvider" # app name
-  module: "dubbo-go greeter  client"
-  version: "0.0.1"
-  environment: "dev"
-
-# registry config
-registries:
-  "demoZk":
-    protocol: "zookeeper"
-    timeout: "3s"
-    address: "127.0.0.1:2181"
-
-# reference config
-references:
-  "greeterImpl": # reference key must match service key in server.yml
-    registry: "demoZk"
-    protocol: "tri"
-    interface: "org.apache.dubbo.IGreeter"
