+++ example/dubbo/go-client/conf/dubbogo.yml
@@ -1,0 +1,13 @@
+dubbo:
+  registries:
+    demoZK:
+      protocol: zookeeper
+      timeout: 3s
+      address: 127.0.0.1:2181
+  consumer:
+    registry:
+      - demoZK
+    references:
+      greeterImpl:
+        protocol: tri
+        interface: org.apache.dubbo.IGreeter # must be compatible with grpc or dubbo-java
