+++ example/dubbo/go-server/conf/dubbogo.yml
@@ -1,0 +1,24 @@
+dubbo:
+  registries:
+    demoZK:
+      protocol: zookeeper
+      timeout: 3s
+      address: 127.0.0.1:2181
+  protocols:
+    triple:
+      name: tri
+      port: 20000
+  provider:
+    registry:
+      - demoZK
+    services:
+      greeterImpl:
+        protocol: triple
+        interface: org.apache.dubbo.IGreeter # must be compatible with grpc or dubbo-java
+  consumer:
+    registry:
+      - demoZK
+    references:
+      greeterImpl:
+        protocol: tri
+        interface: org.apache.dubbo.IGreeter # must be compatible with grpc or dubbo-java
