--- example/dubbo/go-server/conf/server.yml
@@ -1,28 +1,0 @@
-# dubbo-go3 server yaml configure file
-# application config (not necessary)
-application:
-  organization: "dubbo.io"
-  name: "GreeterProvider"
-  module: "dubbo-go3.0 server"
-  version: "0.0.1"
-  environment: "dev"
-
-# registry config
-registries:
-  "demoZK":
-    protocol: "zookeeper"
-    timeout: "3s"
-    address: "127.0.0.1:2181"
-
-# service config
-services:
-  "greeterImpl": # service key must match reference key in client.yml
-    registry: "demoZK"
-    protocol: "tri"
-    interface: "org.apache.dubbo.IGreeter" # must be compatible with grpc or dubbo-java
-
-# protocol config
-protocols:
-  "tri":
-    name: "tri"
-    port: 20001
