+++ example/dubbo/go-server/tests/main_test.go
@@ -7,22 +7,15 @@ import (
 )
 
 import (
-	_ "dubbo.apache.org/dubbo-go/v3/cluster/cluster_impl"
-	_ "dubbo.apache.org/dubbo-go/v3/cluster/loadbalance"
-	_ "dubbo.apache.org/dubbo-go/v3/common/proxy/proxy_factory"
 	"dubbo.apache.org/dubbo-go/v3/config"
-	_ "dubbo.apache.org/dubbo-go/v3/filter/filter_impl"
-	_ "dubbo.apache.org/dubbo-go/v3/protocol/dubbo3"
-	_ "dubbo.apache.org/dubbo-go/v3/protocol/grpc"
-	_ "dubbo.apache.org/dubbo-go/v3/registry/protocol"
-	_ "dubbo.apache.org/dubbo-go/v3/registry/zookeeper"
+	_ "dubbo.apache.org/dubbo-go/v3/imports"
 )
 
 import (
-	client "github.com/dubbogo/triple/example/dubbo/go-client/pkg"
+	"github.com/dubbogo/triple/example/dubbo/proto"
 )
 
-var greeterProvider = new(client.GreeterClientImpl)
+var greeterProvider = new(proto.GreeterClientImpl)
 
 func init() {
 	config.SetConsumerService(greeterProvider)
