+++ example/dubbo/go-server/tests/sayhello_test.go
@@ -2,6 +2,7 @@ package tests
 
 import (
 	"context"
+
 	"testing"
 )
 
@@ -10,19 +11,19 @@ import (
 )
 
 import (
-	client "github.com/dubbogo/triple/example/dubbo/go-client/pkg"
-	tripleConstant "github.com/dubbogo/triple/pkg/common/constant"
+	"github.com/dubbogo/triple/example/dubbo/proto"
 )
 
 func TestSayHello(t *testing.T) {
 	ctx := context.Background()
-	ctx = context.WithValue(ctx, tripleConstant.TripleCtxKey(tripleConstant.TripleRequestID), "triple-request-id-demo")
 
-	req := client.HelloRequest{
+	//ctx = context.WithValue(ctx, tripleConstant.TripleCtxKey(tripleConstant.TripleRequestID), "triple-request-id-demo")
+	//ctx = context.WithValue(ctx, tripleConstant.TripleAttachement, "triple-request-id-demo")
+
+	req := proto.HelloRequest{
 		Name: "laurence",
 	}
-	user := client.User{}
-	err := greeterProvider.SayHello(ctx, &req, &user)
+	user, err := greeterProvider.SayHello(ctx, &req)
 	assert.Nil(t, err)
 
 	assert.Equal(t, "laurence", user.Name)
