+++ example/dubbo/java-server/run.sh
