+++ example/start_integrate_test.sh
@@ -14,7 +14,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-array+=("dubbo/go-server")
+array=("dubbo/go-server")
 array+=("http2/server")
