+++ example/waiting_launch.sh
